Added nth_iterator, which starts at begin(rng) + n.
nth_iterator is necessary to let drop advance begin(rng) by
distance(rng) when Rng is a CommonRange, but isn't random-access. It
also adds an optimisation that lets us take from the back of a range if
we're closer to the end than the front (and the CommonRange is also
Bidirectional).