adds `negate` function object, fixes minus requirements, adjusts whitespace
The ranges-equivalent of std::negate was forgotten in the previous
commit. This commit is to rectify that mistake.

The requirement for a type to have a unary operator- is now under the
jurisdiction of the implementation-detail concept __negatable, which is
refined by __differenceable_with. A slight bug was identified: -t should
not be required to return the same type as T, but rather a type in
common with T.

Finally, there were a few whitespace issues that were noticed. They've
been cleaned up.