adds ranges::ext::partial_sum
The design of this algorithm is slightly different to the one found in
range-v3, as it has a slightly weaker requirement. The range-v3 cousin
requires the binary operation to be a semigroup over the projection of
I; but the author couldn't see a reason for partial_sum to require
associativity. As such, the implementation here only requires that the
binary operation model a magma over the projection of I.