adds a test for numeric concepts and applies patches
This is probably some of the most questionable code that I've written,
but is necessary because C++20 doesn't support concept templates. The
test attached demonstrates that Magma, Semigroup, and Monoid work as
expected for lvalues, rvalues, and references. It's likely incomplete.

The tests identified that there were a few errors embedded into the
design of `two_sided_identity`, and promted a small patch to some code
that GCC permits, but the IS does not.